Feature : New Method For Implementation

+++ Account.Application.Library/BaseService/IBaseQueries.cs
@@ -1,14 +1,37 @@
 ﻿using Account.Application.Library.Models.Controls;
-using System.Data;
 
 namespace Account.Application.Library.BaseService
 {
     public interface IBaseQueries
     {
+        /// <summary>
+        /// دریافت تعداد رکورد ها
+        /// </summary>
+        /// <returns></returns>
         string GetCount();
+        /// <summary>
+        /// نمایش همه رکورد ها
+        /// </summary>
+        /// <param name="paging"></param>
+        /// <returns></returns>
         string ShowAll(string paging);
+        /// <summary>
+        /// جستجو در میان رکورد ها
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         string Search(string value);
+        /// <summary>
+        /// جستجو در بازه از تاریخ تا تاریخ
+        /// </summary>
+        /// <param name="from"></param>
+        /// <param name="to"></param>
+        /// <returns></returns>
         string ShowFromTo(string from, string to);
-        IEnumerable<KeyValue<long>> TitleValue();
+        /// <summary>
+        /// دریافت عنوان و مقدار
+        /// </summary>
+        /// <returns></returns>
+        IEnumerable<KeyValue<byte>> TitleValue();
     }
 }
